Fully move to one file for datasets page and migrate to bootstrap CDN
