+++ _includes/footer.html
@@ -1,4 +1,2 @@
-<!-- general behavior -->
-<script type="module" src="{{ site.baseurl }}/scripts/dist/index.js"></script>
-
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous"></script>
 <footer></footer>
