+++ datasets.html
@@ -25,7 +25,7 @@ permalink: /datasets/
     <div class="list-group js-hidden">
       {% for organization in site.organizations %}
       {% assign count = site.datasets | where: "organization", organization.title | size %}
-      <a href="{{ site.baseurl}}{{ organization.url }}" class="list-group-item list-group-item-action d-flex justify-content-between align-items-center">
+      <a href="{{ site.baseurl }}{{ organization.url }}" class="list-group-item list-group-item-action d-flex justify-content-between align-items-center">
         {{ organization.title }}
         <span class="badge bg-primary rounded-pill">{{ count }}</span>
       </a>
@@ -46,3 +46,6 @@ permalink: /datasets/
     </div>
   </div>
 </div>
+
+<!-- Load the javascript for search and category/org filtering -->
+<script src="{{ site.baseurl }}/scripts/datasets.js"></script>
