--- package.json
@@ -1,29 +1,0 @@
-{
-  "name": "jkan",
-  "description": "A lightweight, backend-free open data portal, powered by Jekyll",
-  "scripts": {
-    "watch": "parcel watch scripts/src/index.js --dist-dir _site/scripts/dist",
-    "build": "parcel build scripts/src/index.js --dist-dir scripts/dist"
-  },
-  "targets": {
-    "modern": {
-      "engines": {
-        "browsers": "defaults"
-      }
-    }
-  },
-  "license": "MIT",
-  "private": true,
-  "bugs": {
-    "url": "https://github.com/azavea/opendataphilly-jkan/issues"
-  },
-  "homepage": "https://github.com/azavea/opendataphilly-jkan/",
-  "devDependencies": {
-    "core-js": "^3.45.1",
-    "parcel": "^2.15.4",
-    "process": "^0.11.10"
-  },
-  "dependencies": {
-    "bootstrap": "^5.3.8"
-  }
-}
