--- scripts/src/templates/dataset-item.js
@@ -1,6 +1,0 @@
-export default (data) => (
-`<dataset>
-  <div class="browse-cat"><a href="${data.url}">${data.title}</a></div>
-  ${data.notes || ''}
-</dataset>`
-)
